requirements.txt
google-api-python-client
termcolor
google-auth-httplib2 
google-auth-oauthlib